[UPDATE]
[s] - 0230
[e] - 0700
[t] - 0430

updated the materials for the battery and mucked around with reflections. NOTE: reflections suck... hard. Abandoning this quirry for the time being, will likely return to it at a later date, once it becomes neccessary. Most options require a plane to handle reflections, so, things like windows and what have you will have a particular shader on them and some scripted cameras/probes in order to work properly.